backend/api_gateway/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
slowapi==0.1.9
pyyaml==6.0.1




